Make tool_consumers resource private
This change adds a devise authentication to all the actions inside the tool_consumer controller.
The effect of this results in that now devise won't allow anyone that's not logged as admin to create or see TC pages.
Also wrapped all the page inside a container-fluid div to add margins.