Add 'ToolConsumer' select field to 'Timetable' form
This change adds a selection field generated with Rails, uses a collection provided by the controller, and to make this work it necessary to add the 'tool_consumer_id' parameter to the list of allowed params.
Also includes a link to timetables at the dashboard and removes an unnecesary '=>' at home_controller.