Add Heroku pipeline and auto-deploy screenshots
At the current moment when a commit is made to the heroku/dev branch is automatically deployed at the 'attnd-ctrl-dev' app.
And when a pull request is made at the entire repository its possible to deploy it to a reviewing app at the same pipeline to test it in an environment.