Initial navbar iteration
This change only pastes a navbar from a bootstrap template on top of the previous 'provisional navigation ul'.
This is just a start point, the objective of this branch is to implement a functional navbar, with a logo, user connected dropdown, and a 'decent' looking style.
Also fixed some tabs.