Fix problem, different DB config between Docker and local
This change fixes the problem that was gonna happen when merging this to 'master'.
To run the docker-compose, the database host must be set to 'db', but to run the local environment it must be '127.0.0.1'.
To fix this a new file named 'database.docker.yml' is created, this will contain the docker DB config, and it will replace the 'database.yml' file used by default.