Add TimetableUnits using FullCalendar
This change allows the creation of Timetable Units using the 'select' function of FullCalendar.
That function calls a Js method that displays a modal with the data of the selection and the name and description field.
To make it work, Turbolinks is disabled at the 'Timetable#show' page.
Attached gif.